+++ event-details/arrival-tracking/index.ts
@@ -1,5 +1,5 @@
 export * from "./Tracker"
-export * from "./WHY_MIO_WHY_WOULD_YOU_DO_THIS_TO_ME_YOU_SAID_YOU_WOULD_BE_BY_MY_SIDE_MIO_ALL_I_DID_I_DID_FOR_YOU"
+export * from "./Models"
 export * from "./Geofencing"
 export * from "./UpcomingArrivals"
 export * from "./Refresh"
